requirements.txt
discord.py==1.3.4
pyNacl==1.3.0
pandas
dnspython==1.16.0
async-timeout==3.0.1
requests